+++ src/Web/ClientApp/src/entities/InputTypes.tsx
@@ -1,0 +1,12 @@
+// export const useInputTypes = [
+//   { value: 0, label: "テキスト入力" },
+//   { value: 1, label: "数値入力" },
+//   { value: 2, label: "項目選択" },
+//   { value: 5, label: "日付入力" },
+// ];
+export const useInputTypes = [
+  { value: 1, label: "テキスト入力" },
+  { value: 2, label: "数値入力" },
+  { value: 3, label: "項目選択" },
+  { value: 6, label: "日付入力" },
+];
