+++ src/Web/ClientApp/src/entities/index.ts
@@ -4,3 +4,5 @@ export * from './Equipment';
 export * from './InspectionSheet';
 export * from './InspectionItemReducer';
 export * from './InspectionSheetReducer';
+export * from './ItemType';
+export * from './InputTypes';
